+++ src/builder.ts
@@ -1,18 +1,12 @@
 import { Database, Table } from './database'
 import { BooleanColumn, Column } from './columns'
-import { Condition } from './models'
+import { Condition } from './models/Condition'
 import { Binder, BinderStore } from './binder'
 import { ASTERISK, Distinct, All } from './singletoneConstants'
-import {
-  Step,
-  SelectStep,
-  FromStep,
-  Parenthesis,
-  LogicalOperator,
-  SelectItem,
-  PrimitiveType,
-  RootStep,
-} from './steps/steps'
+import { SelectStep, FromStep, RootStep } from './steps/stepInterfaces'
+import { Step, PrimitiveType, SelectItem } from './steps/Step'
+import { LogicalOperator } from './operators'
+import { Parenthesis } from './steps/BaseStep'
 import { OrderByItemInfo } from './orderBy'
 import { SelectItemInfo } from './select'
 import { BuilderOption, fillUndefinedOptionsWithDefault } from './option'