+++ src/steps/BaseStep.ts
@@ -1,9 +1,16 @@
 import { BuilderData } from '../builder'
-import { Condition, Expression, PostgresBinder } from '../models'
+import { Condition } from '../models/Condition'
+import { PostgresBinder } from '../models/types'
+import { Expression } from '../models/Expression'
 import { BooleanColumn } from '../columns'
-import { LogicalOperator, Parenthesis } from './steps'
+import { LogicalOperator } from '../operators'
 
-export class BaseStep {
+export enum Parenthesis {
+  Open = '(',
+  Close = ')',
+}
+
+export abstract class BaseStep {
   constructor(protected data: BuilderData) {}
 
   public getSQL(): string {
@@ -22,7 +29,7 @@ export class BaseStep {
 
     if (this.data.selectItemInfos.length > 0) {
       const selectPartsString = this.data.selectItemInfos.map(it => {
-        return it.getStmt({ binderStore: this.data.binderStore })
+        return it.getStmt(this.data)
       })
       result += ` ${selectPartsString.join(', ')}`
     }
@@ -63,7 +70,7 @@ export class BaseStep {
 
     if (this.data.orderByItemInfos.length > 0) {
       const orderByPartsString = this.data.orderByItemInfos.map(it => {
-        return it.getStmt({ binderStore: this.data.binderStore })
+        return it.getStmt(this.data)
       })
       result += ` ORDER BY ${orderByPartsString.join(', ')}`
     }