+++ src/steps/Step.ts
@@ -1,4 +1,5 @@
-import { Condition, Expression } from '../models'
+import { Condition } from '../models/Condition'
+import { Expression } from '../models/Expression'
 import { Column } from '../columns'
 import { Table } from '../database'
 import { ColumnNotFoundError, TableNotFoundError } from '../errors'
@@ -18,6 +19,8 @@ import { Binder } from '../binder'
 import { BaseStep } from './BaseStep'
 import { WhereStep } from './WhereStep'
 import { HavingStep } from './HavingStep'
+import { RootStep, SelectStep, FromStep, GroupByStep, OrderByStep, LimitStep, OffsetStep } from './stepInterfaces'
+import { LogicalOperator } from '../operators'
 
 export type ColumnLike = Column|Expression
 export type PrimitiveType = null|boolean|number|string
@@ -199,137 +202,4 @@ export class Step extends BaseStep implements RootStep, SelectStep, FromStep, Gr
       }
     }
   }
-}
-
-//@formatter:off
-export interface RootStep extends BaseStep {
-  select(...items: (SelectItemInfo|SelectItem|PrimitiveType)[]): SelectStep
-  selectDistinct(...items: (SelectItemInfo|SelectItem|PrimitiveType)[]): SelectStep
-  selectAll(...items: (SelectItemInfo|SelectItem|PrimitiveType)[]): SelectStep
-}
-
-export interface SelectStep extends BaseStep {
-  from(table: Table): FromStep
-}
-
-export interface FromStep extends BaseStep {
-  where(condition: Condition): WhereStep
-  where(left: Condition, operator: LogicalOperator, right: Condition): WhereStep
-  where(left: Condition, operator1: LogicalOperator, middle: Condition, operator2: LogicalOperator, right: Condition): WhereStep
-
-  groupBy(...groupByItems: Column[]): GroupByStep
-  orderBy(...orderByItems: OrderByArgsElement[]): OrderByStep
-  limit(n: null|number|All): LimitStep
-  limit$(n: null|number): LimitStep
-  offset(n: number): OffsetStep
-  offset$(n: number): OffsetStep
-}
-
-export interface WhereAndStep extends BaseStep {
-  and(condition: Condition): WhereAndStep
-  and(left: Condition, operator: LogicalOperator, right: Condition): WhereAndStep
-  and(left: Condition, operator1: LogicalOperator, middle: Condition, operator2: LogicalOperator, right: Condition): WhereAndStep
-
-  or(condition: Condition): WhereOrStep
-  or(left: Condition, operator: LogicalOperator, right: Condition): WhereOrStep
-  or(left: Condition, operator1: LogicalOperator, middle: Condition, operator2: LogicalOperator, right: Condition): WhereOrStep
-
-  groupBy(...groupByItems: Column[]): GroupByStep
-  orderBy(...orderByItems: OrderByArgsElement[]): OrderByStep
-  limit(n: null|number|All): LimitStep
-  limit$(n: null|number): LimitStep
-  offset(n: number): OffsetStep
-  offset$(n: number): OffsetStep
-}
-
-export interface WhereOrStep extends BaseStep {
-  or(condition: Condition): WhereOrStep
-  or(left: Condition, operator: LogicalOperator, right: Condition): WhereOrStep
-  or(left: Condition, operator1: LogicalOperator, middle: Condition, operator2: LogicalOperator, right: Condition): WhereOrStep
-
-  and(condition: Condition): WhereAndStep
-  and(left: Condition, operator: LogicalOperator, right: Condition): WhereAndStep
-  and(left: Condition, operator1: LogicalOperator, middle: Condition, operator2: LogicalOperator, right: Condition): WhereAndStep
-
-  groupBy(...groupByItems: Column[]): GroupByStep
-  orderBy(...orderByItems: OrderByArgsElement[]): OrderByStep
-  limit(n: null|number|All): LimitStep
-  limit$(n: null|number): LimitStep
-  offset(n: number): OffsetStep
-  offset$(n: number): OffsetStep
-}
-
-export interface GroupByStep extends BaseStep {
-  having(condition: Condition): HavingStep
-  having(left: Condition, operator: LogicalOperator, right: Condition): HavingStep
-  having(left: Condition, operator1: LogicalOperator, middle: Condition, operator2: LogicalOperator, right: Condition): HavingStep
-
-  orderBy(...orderByItems: OrderByArgsElement[]): OrderByStep
-
-  limit(n: null|number|All): LimitStep
-  limit$(n: null|number): LimitStep
-
-  offset(n: number): OffsetStep
-  offset$(n: number): OffsetStep
-}
-
-export interface HavingOrStep extends BaseStep {
-  or(condition: Condition): HavingOrStep
-  or(left: Condition, operator: LogicalOperator, right: Condition): HavingOrStep
-  or(left: Condition, operator1: LogicalOperator, middle: Condition, operator2: LogicalOperator, right: Condition): HavingOrStep
-
-  and(condition: Condition): HavingAndStep
-  and(left: Condition, operator: LogicalOperator, right: Condition): HavingOrStep
-  and(left: Condition, operator1: LogicalOperator, middle: Condition, operator2: LogicalOperator, right: Condition): HavingOrStep
-
-  orderBy(...orderByItems: OrderByArgsElement[]): OrderByStep
-
-  limit(n: null|number|All): LimitStep
-  limit$(n: null|number): LimitStep
-
-  offset(n: number): OffsetStep
-  offset$(n: number): OffsetStep
-}
-
-export interface HavingAndStep extends BaseStep {
-  and(condition: Condition): HavingAndStep
-  and(left: Condition, operator: LogicalOperator, right: Condition): HavingOrStep
-  and(left: Condition, operator1: LogicalOperator, middle: Condition, operator2: LogicalOperator, right: Condition): HavingOrStep
-
-  or(condition: Condition): HavingOrStep
-  or(left: Condition, operator: LogicalOperator, right: Condition): HavingOrStep
-  or(left: Condition, operator1: LogicalOperator, middle: Condition, operator2: LogicalOperator, right: Condition): HavingOrStep
-
-  orderBy(...orderByItems: OrderByArgsElement[]): OrderByStep
-
-  limit(n: null|number|All): LimitStep
-  limit$(n: null|number): LimitStep
-
-  offset(n: number): OffsetStep
-  offset$(n: number): OffsetStep
-}
-
-export interface OrderByStep extends BaseStep {
-  limit(n: null|number|All): LimitStep
-  limit$(n: null|number): LimitStep
-  offset(n: number): OffsetStep
-  offset$(n: number): OffsetStep
-}
-
-export interface LimitStep extends BaseStep {
-  offset(n: number): OffsetStep
-  offset$(n: number): OffsetStep
-}
-
-export interface OffsetStep extends BaseStep {}
-//@formatter:on
-
-export enum LogicalOperator {
-  AND = 'AND',
-  OR = 'OR',
-}
-
-export enum Parenthesis {
-  Open = '(',
-  Close = ')',
 }