+++ test/HavingStep.test.ts
@@ -1,14 +1,15 @@
-import { Builder, LogicalOperator } from '../src'
+import { Builder, LogicalOperator, f } from '../src'
 import { database } from './database'
 
 //Alias
 const AND = LogicalOperator.AND
 const OR = LogicalOperator.OR
 const table = database.s.public.t.testTable
-const column1 = database.s.public.t.testTable.c.column1
-const column2 = database.s.public.t.testTable.c.column2
-const column3 = database.s.public.t.testTable.c.column3
-const column4 = database.s.public.t.testTable.c.column4
+const col1 = table.c.col1
+const col2 = table.c.col2
+const col3 = table.c.col3
+const col4 = table.c.col4
+const col5 = table.c.col5
 
 describe('test groupBy Step', () => {
   const sql = new Builder(database)
@@ -16,10 +17,10 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING "col1" = \'a\';]', () => {
     const actual = sql
-      .select(column1)
+      .select(col1)
       .from(table)
-      .groupBy(column1)
-      .having(column1.eq('a'))
+      .groupBy(col1)
+      .having(col1.eq('a'))
       .getSQL()
 
     expect(actual).toEqual('SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING "col1" = \'a\';')
@@ -27,10 +28,10 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING ( "col1" = \'a\' AND "col2" = \'b\' );]', () => {
     const actual = sql
-      .select(column1)
+      .select(col1)
       .from(table)
-      .groupBy(column1)
-      .having(column1.eq('a'), AND, column2.eq('b'))
+      .groupBy(col1)
+      .having(col1.eq('a'), AND, col2.eq('b'))
       .getSQL()
 
     expect(actual).toEqual('SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING ( "col1" = \'a\' AND "col2" = \'b\' );')
@@ -38,10 +39,10 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING ( "col1" = \'a\' AND "col2" = \'b\' OR "col3" = \'c\' );]', () => {
     const actual = sql
-      .select(column1)
+      .select(col1)
       .from(table)
-      .groupBy(column1)
-      .having(column1.eq('a'), AND, column2.eq('b'), OR, column3.eq('c'))
+      .groupBy(col1)
+      .having(col1.eq('a'), AND, col2.eq('b'), OR, col3.eq('c'))
       .getSQL()
 
     expect(actual).toEqual('SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING ( "col1" = \'a\' AND "col2" = \'b\' OR "col3" = \'c\' );')
@@ -49,11 +50,11 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING "col1" = \'a\' AND "col2" = \'b\';]', () => {
     const actual = sql
-      .select(column1)
+      .select(col1)
       .from(table)
-      .groupBy(column1)
-      .having(column1.eq('a'))
-      .and(column2.eq('b'))
+      .groupBy(col1)
+      .having(col1.eq('a'))
+      .and(col2.eq('b'))
       .getSQL()
 
     expect(actual).toEqual('SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING "col1" = \'a\' AND "col2" = \'b\';')
@@ -61,11 +62,11 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING "col1" = \'a\' AND ( "col2" = \'b\' OR "col3" = \'c\' );]', () => {
     const actual = sql
-      .select(column1)
+      .select(col1)
       .from(table)
-      .groupBy(column1)
-      .having(column1.eq('a'))
-      .and(column2.eq('b'), OR, column3.eq('c'))
+      .groupBy(col1)
+      .having(col1.eq('a'))
+      .and(col2.eq('b'), OR, col3.eq('c'))
       .getSQL()
 
     expect(actual).toEqual('SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING "col1" = \'a\' AND ( "col2" = \'b\' OR "col3" = \'c\' );')
@@ -73,11 +74,11 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING "col1" = \'a\' AND ( "col2" = \'b\' OR "col3" = \'c\' AND "col4" = 4 );]', () => {
     const actual = sql
-      .select(column1)
+      .select(col1)
       .from(table)
-      .groupBy(column1)
-      .having(column1.eq('a'))
-      .and(column2.eq('b'), OR, column3.eq('c'), AND, column4.eq(4))
+      .groupBy(col1)
+      .having(col1.eq('a'))
+      .and(col2.eq('b'), OR, col3.eq('c'), AND, col4.eq(4))
       .getSQL()
 
     expect(actual).toEqual('SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING "col1" = \'a\' AND ( "col2" = \'b\' OR "col3" = \'c\' AND "col4" = 4 );')
@@ -85,11 +86,11 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING "col1" = \'a\' OR ( "col2" = \'b\' OR "col3" = \'c\' AND "col4" = 4 );]', () => {
     const actual = sql
-      .select(column1)
+      .select(col1)
       .from(table)
-      .groupBy(column1)
-      .having(column1.eq('a'))
-      .or(column2.eq('b'), OR, column3.eq('c'), AND, column4.eq(4))
+      .groupBy(col1)
+      .having(col1.eq('a'))
+      .or(col2.eq('b'), OR, col3.eq('c'), AND, col4.eq(4))
       .getSQL()
 
     expect(actual).toEqual('SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING "col1" = \'a\' OR ( "col2" = \'b\' OR "col3" = \'c\' AND "col4" = 4 );')
@@ -97,14 +98,79 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING "col1" = \'a\' AND "col2" = \'b\' OR "col3" = \'c\';]', () => {
     const actual = sql
-      .select(column1)
+      .select(col1)
       .from(table)
-      .groupBy(column1)
-      .having(column1.eq('a'))
-      .and(column2.eq('b'))
-      .or(column3.eq('c'))
+      .groupBy(col1)
+      .having(col1.eq('a'))
+      .and(col2.eq('b'))
+      .or(col3.eq('c'))
       .getSQL()
 
     expect(actual).toEqual('SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING "col1" = \'a\' AND "col2" = \'b\' OR "col3" = \'c\';')
+  })
+
+  it('Produces [SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING SUM("col4") = 4;]', () => {
+    const actual = sql
+      .select(col1)
+      .from(table)
+      .groupBy(col1)
+      .having(f.sum(col4).eq(4))
+      .getSQL()
+
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING SUM("col4") = 4;')
+  })
+
+  it('Produces [SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING SUM("col5") = 5;]', () => {
+    const actual = sql
+      .select(col1)
+      .from(table)
+      .groupBy(col1)
+      .having(col5.sum.eq(5))
+      .getSQL()
+
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING SUM("col5") = 5;')
+  })
+
+  it('Produces [SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING SUM("col4") = $1;]', () => {
+    const actual = sql
+      .select(col1)
+      .from(table)
+      .groupBy(col1)
+      .having(col4.sum.eq$(4))
+      .getBinds()
+
+    const expected = {
+      sql: 'SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING SUM("col4") = $1;',
+      values: [4],
+    }
+
+    expect(actual).toEqual(expected)
+  })
+
+  it('Produces [SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING SUM("col5") <> 5;]', () => {
+    const actual = sql
+      .select(col1)
+      .from(table)
+      .groupBy(col1)
+      .having(col5.count.ne(5))
+      .getSQL()
+
+    expect(actual).toEqual('SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING COUNT("col5") <> 5;')
+  })
+
+  it('Produces [SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING AVG("col4") > $1;]', () => {
+    const actual = sql
+      .select(col1)
+      .from(table)
+      .groupBy(col1)
+      .having(col4.avg.gt$(4))
+      .getBinds()
+
+    const expected = {
+      sql: 'SELECT "col1" FROM "testTable" GROUP BY "col1" HAVING AVG("col4") > $1;',
+      values: [4],
+    }
+
+    expect(actual).toEqual(expected)
   })
 })