+++ test/groupByStep.test.ts
@@ -4,12 +4,12 @@ import { database } from './database'
 //Alias
 const ADD = ArithmeticOperator.ADD
 const table = database.s.public.t.testTable
-const column1 = database.s.public.t.testTable.c.column1
-const column2 = database.s.public.t.testTable.c.column2
-const column3 = database.s.public.t.testTable.c.column3
-const column4 = database.s.public.t.testTable.c.column4
-const column5 = database.s.public.t.testTable.c.column5
-const column6 = database.s.public.t.testTable.c.column6
+const col1 = database.s.public.t.testTable.c.col1
+const col2 = database.s.public.t.testTable.c.col2
+const col3 = database.s.public.t.testTable.c.col3
+const col4 = database.s.public.t.testTable.c.col4
+const col5 = database.s.public.t.testTable.c.col5
+const col6 = database.s.public.t.testTable.c.col6
 
 describe('test groupBy Step', () => {
   const sql = new Builder(database)
@@ -17,9 +17,9 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT "col1" FROM "testTable" GROUP BY "col1";]', () => {
     const actual = sql
-      .select(column1)
+      .select(col1)
       .from(table)
-      .groupBy(column1)
+      .groupBy(col1)
       .getSQL()
 
     expect(actual).toEqual('SELECT "col1" FROM "testTable" GROUP BY "col1";')
@@ -27,9 +27,9 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT "col1", "col2" FROM "testTable" GROUP BY "col1", "col2";]', () => {
     const actual = sql
-      .select(column1, column2)
+      .select(col1, col2)
       .from(table)
-      .groupBy(column1, column2)
+      .groupBy(col1, col2)
       .getSQL()
 
     expect(actual).toEqual('SELECT "col1", "col2" FROM "testTable" GROUP BY "col1", "col2";')
@@ -37,10 +37,10 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT "col1" FROM "testTable" WHERE "col2" = \'a\' GROUP BY "col1";]', () => {
     const actual = sql
-      .select(column1)
+      .select(col1)
       .from(table)
-      .where(column2.eq('a'))
-      .groupBy(column1)
+      .where(col2.eq('a'))
+      .groupBy(col1)
       .getSQL()
 
     expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col2" = \'a\' GROUP BY "col1";')
@@ -48,11 +48,11 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT "col1" FROM "testTable" WHERE "col2" = \'a\' AND "col3" = \'b\' GROUP BY "col1";]', () => {
     const actual = sql
-      .select(column1)
+      .select(col1)
       .from(table)
-      .where(column2.eq('a'))
-      .and(column3.eq('b'))
-      .groupBy(column1)
+      .where(col2.eq('a'))
+      .and(col3.eq('b'))
+      .groupBy(col1)
       .getSQL()
 
     expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col2" = \'a\' AND "col3" = \'b\' GROUP BY "col1";')
@@ -60,11 +60,11 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT "col1" FROM "testTable" WHERE "col2" = \'a\' OR "col3" = \'b\' GROUP BY "col1";]', () => {
     const actual = sql
-      .select(column1)
+      .select(col1)
       .from(table)
-      .where(column2.eq('a'))
-      .or(column3.eq('b'))
-      .groupBy(column1)
+      .where(col2.eq('a'))
+      .or(col3.eq('b'))
+      .groupBy(col1)
       .getSQL()
 
     expect(actual).toEqual('SELECT "col1" FROM "testTable" WHERE "col2" = \'a\' OR "col3" = \'b\' GROUP BY "col1";')
@@ -72,9 +72,9 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT SUM("col4"), SUM(1) FROM "testTable" GROUP BY "col2";]', () => {
     const actual = sql
-      .select(f.sum(column4), f.sum(1))
+      .select(f.sum(col4), f.sum(1))
       .from(table)
-      .groupBy(column2)
+      .groupBy(col2)
       .getSQL()
 
     expect(actual).toEqual('SELECT SUM("col4"), SUM(1) FROM "testTable" GROUP BY "col2";')
@@ -82,9 +82,9 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT SUM("col4"), AVG("col5") FROM "testTable" GROUP BY "col2";]', () => {
     const actual = sql
-      .select(f.sum(column4), f.avg(column5))
+      .select(f.sum(col4), f.avg(col5))
       .from(table)
-      .groupBy(column2)
+      .groupBy(col2)
       .getSQL()
 
     expect(actual).toEqual('SELECT SUM("col4"), AVG("col5") FROM "testTable" GROUP BY "col2";')
@@ -92,9 +92,9 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT COUNT("col4"), MAX("col5"), MIN("col6") FROM "testTable" GROUP BY "col2";]', () => {
     const actual = sql
-      .select(f.count(column4), f.max(column5), f.min(column6))
+      .select(f.count(col4), f.max(col5), f.min(col6))
       .from(table)
-      .groupBy(column2)
+      .groupBy(col2)
       .getSQL()
 
     expect(actual).toEqual('SELECT COUNT("col4"), MAX("col5"), MIN("col6") FROM "testTable" GROUP BY "col2";')
@@ -102,9 +102,9 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT SUM("col4"), AVG("col4"), COUNT("col4"), MAX("col4"), MIN("col4") FROM "testTable" GROUP BY "col2";]', () => {
     const actual = sql
-      .select(column4.sum, column4.avg, column4.count, column4.max, column4.min)
+      .select(col4.sum, col4.avg, col4.count, col4.max, col4.min)
       .from(table)
-      .groupBy(column2)
+      .groupBy(col2)
       .getSQL()
 
     expect(actual).toEqual('SELECT SUM("col4"), AVG("col4"), COUNT("col4"), MAX("col4"), MIN("col4") FROM "testTable" GROUP BY "col2";')
@@ -113,12 +113,12 @@ describe('test groupBy Step', () => {
   it('Produces [SELECT SUM("col4") AS "SUM_OF_COL4", SUM(1) AS "SUM_OF_1", SUM(col5) AS "SUM_OF_COL5" FROM "testTable" GROUP BY "col1";]', () => {
     const actual = sql
       .select(
-        f.sum(column4).as('SUM_OF_COL4'),
+        f.sum(col4).as('SUM_OF_COL4'),
         f.sum(1).as('SUM_OF_1'),
-        column5.sum.as('SUM_OF_COL5'),
+        col5.sum.as('SUM_OF_COL5'),
       )
       .from(table)
-      .groupBy(column2)
+      .groupBy(col2)
       .getSQL()
 
     expect(actual).toEqual('SELECT SUM("col4") AS "SUM_OF_COL4", SUM(1) AS "SUM_OF_1", SUM("col5") AS "SUM_OF_COL5" FROM "testTable" GROUP BY "col2";')
@@ -126,9 +126,9 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT SUM("col4" + "col5") FROM "testTable" GROUP BY "col1";]', () => {
     const actual = sql
-      .select(f.sum(e(column4, ADD, column5)))
+      .select(f.sum(e(col4, ADD, col5)))
       .from(table)
-      .groupBy(column1)
+      .groupBy(col1)
       .getSQL()
 
     expect(actual).toEqual('SELECT SUM("col4" + "col5") FROM "testTable" GROUP BY "col1";')
@@ -136,9 +136,9 @@ describe('test groupBy Step', () => {
 
   it('Produces [SELECT SUM("col4" + 1) FROM "testTable" GROUP BY "col1";]', () => {
     const actual = sql
-      .select(f.sum(e(column4, ADD, 1)))
+      .select(f.sum(e(col4, ADD, 1)))
       .from(table)
-      .groupBy(column1)
+      .groupBy(col1)
       .getSQL()
 
     expect(actual).toEqual('SELECT SUM("col4" + 1) FROM "testTable" GROUP BY "col1";')