+++ test/orderByStep.test.ts
@@ -11,11 +11,11 @@ import {
 import { database } from './database'
 //Alias
 const table = database.s.public.t.testTable
-const column1 = database.s.public.t.testTable.c.column1
-const column2 = database.s.public.t.testTable.c.column2
-const column3 = database.s.public.t.testTable.c.column3
-const column4 = database.s.public.t.testTable.c.column4
-const column5 = database.s.public.t.testTable.c.column5
+const col1 = database.s.public.t.testTable.c.col1
+const col2 = database.s.public.t.testTable.c.col2
+const col3 = database.s.public.t.testTable.c.col3
+const col4 = database.s.public.t.testTable.c.col4
+const col5 = database.s.public.t.testTable.c.col5
 
 describe('test orderBy Step', () => {
   const sql = new Builder(database)
@@ -24,7 +24,7 @@ describe('test orderBy Step', () => {
   it('Produces [SELECT * FROM "testTable" ORDER BY "col1", "col2";]', () => {
     const actual = sql
       .selectAsteriskFrom(table)
-      .orderBy(column1, column2)
+      .orderBy(col1, col2)
       .getSQL()
 
     expect(actual).toEqual('SELECT * FROM "testTable" ORDER BY "col1", "col2";')
@@ -33,7 +33,7 @@ describe('test orderBy Step', () => {
   it('Produces [SELECT * FROM "testTable" ORDER BY ("col4" + "col5");]', () => {
     const actual = sql
       .selectAsteriskFrom(table)
-      .orderBy(e(column4, ArithmeticOperator.ADD, column5))
+      .orderBy(e(col4, ArithmeticOperator.ADD, col5))
       .getSQL()
 
     expect(actual).toEqual('SELECT * FROM "testTable" ORDER BY ("col4" + "col5");')
@@ -42,7 +42,7 @@ describe('test orderBy Step', () => {
   it('Produces [SELECT * FROM "testTable" ORDER BY col1 DESC NULLS FIRST;]', () => {
     const actual = sql
       .selectAsteriskFrom(table)
-      .orderBy(o(column1, DESC, NULLS_FIRST))
+      .orderBy(o(col1, DESC, NULLS_FIRST))
       .getSQL()
 
     expect(actual).toEqual('SELECT * FROM "testTable" ORDER BY "col1" DESC NULLS FIRST;')
@@ -51,7 +51,7 @@ describe('test orderBy Step', () => {
   it('Produces [SELECT * FROM "testTable" ORDER BY col1 ASC;]', () => {
     const actual = sql
       .selectAsteriskFrom(table)
-      .orderBy(o(column1, DESC, NULLS_FIRST))
+      .orderBy(o(col1, DESC, NULLS_FIRST))
       .getSQL()
 
     expect(actual).toEqual('SELECT * FROM "testTable" ORDER BY "col1" DESC NULLS FIRST;')
@@ -60,7 +60,7 @@ describe('test orderBy Step', () => {
   it('Produces [SELECT * FROM "testTable" ORDER BY ("col4" + "col5") DESC;]', () => {
     const actual = sql
       .selectAsteriskFrom(table)
-      .orderBy(o(e(column4, ArithmeticOperator.ADD, column5), DESC))
+      .orderBy(o(e(col4, ArithmeticOperator.ADD, col5), DESC))
       .getSQL()
 
     expect(actual).toEqual('SELECT * FROM "testTable" ORDER BY ("col4" + "col5") DESC;')
@@ -68,7 +68,7 @@ describe('test orderBy Step', () => {
 
   it('Produces [SELECT ("col4" + "col5") AS "Col:4+5" FROM "testTable" ORDER BY "Col:4+5";]', () => {
     const actual = sql
-      .select(e(column4, ArithmeticOperator.ADD, column5).as('Col:4+5'))
+      .select(e(col4, ArithmeticOperator.ADD, col5).as('Col:4+5'))
       .from(table)
       .orderBy('Col:4+5')
       .getSQL()
@@ -79,7 +79,7 @@ describe('test orderBy Step', () => {
   it('Throws error when ORDER BY alias not exist', () => {
     function actual() {
       sql
-        .select(column1.as('A'))
+        .select(col1.as('A'))
         .from(table)
         .orderBy('B')
         .getSQL()
@@ -91,7 +91,7 @@ describe('test orderBy Step', () => {
   it('Produces [SELECT * FROM "testTable" ORDER BY "col1" ASC, "col2" DESC;]', () => {
     const actual = sql
       .selectAsteriskFrom(table)
-      .orderBy(column1.asc, column2.desc)
+      .orderBy(col1.asc, col2.desc)
       .getSQL()
 
     expect(actual).toEqual('SELECT * FROM "testTable" ORDER BY "col1" ASC, "col2" DESC;')
@@ -100,7 +100,7 @@ describe('test orderBy Step', () => {
   it('Produces [SELECT * FROM "testTable" ORDER BY "col1" ASC, "col2" DESC, "col3" NULLS FIRST, "col4" NULLS LAST;]', () => {
     const actual = sql
       .selectAsteriskFrom(table)
-      .orderBy(column1.asc, column2.desc, column3.nullsFirst, column4.nullsLast)
+      .orderBy(col1.asc, col2.desc, col3.nullsFirst, col4.nullsLast)
       .getSQL()
 
     expect(actual).toEqual('SELECT * FROM "testTable" ORDER BY "col1" ASC, "col2" DESC, "col3" NULLS FIRST, "col4" NULLS LAST;')
@@ -109,7 +109,7 @@ describe('test orderBy Step', () => {
   it('Produces [SELECT * FROM "testTable" ORDER BY "col1" ASC NULLS FIRST, "col2" DESC NULLS FIRST, "col3" ASC NULLS LAST, "col4" DESC NULLS LAST;]', () => {
     const actual = sql
       .selectAsteriskFrom(table)
-      .orderBy(column1.ascNullsFirst, column2.descNullsFirst, column3.ascNullsLast, column4.descNullsLast)
+      .orderBy(col1.ascNullsFirst, col2.descNullsFirst, col3.ascNullsLast, col4.descNullsLast)
       .getSQL()
 
     expect(actual).toEqual('SELECT * FROM "testTable" ORDER BY "col1" ASC NULLS FIRST, "col2" DESC NULLS FIRST, "col3" ASC NULLS LAST, "col4" DESC NULLS LAST;')
@@ -117,9 +117,9 @@ describe('test orderBy Step', () => {
 
   it('Produces [SELECT "col1" AS "C1" FROM "testTable" ORDER BY "C1" DESC NULLS FIRST, "col2" ASC NULLS LAST;]', () => {
     const actual = sql
-      .select(column1.as('C1'))
+      .select(col1.as('C1'))
       .from(table)
-      .orderBy('C1', DESC, NULLS_FIRST, column2, ASC, NULLS_LAST)
+      .orderBy('C1', DESC, NULLS_FIRST, col2, ASC, NULLS_LAST)
       .getSQL()
 
     expect(actual).toEqual('SELECT "col1" AS "C1" FROM "testTable" ORDER BY "C1" DESC NULLS FIRST, "col2" ASC NULLS LAST;')
@@ -127,9 +127,9 @@ describe('test orderBy Step', () => {
 
   it('Produces [SELECT "col1" AS "C1" FROM "testTable" ORDER BY "C1" DESC, "col2" NULLS LAST;]', () => {
     const actual = sql
-      .select(column1.as('C1'))
+      .select(col1.as('C1'))
       .from(table)
-      .orderBy('C1', DESC, column2, NULLS_LAST)
+      .orderBy('C1', DESC, col2, NULLS_LAST)
       .getSQL()
 
     expect(actual).toEqual('SELECT "col1" AS "C1" FROM "testTable" ORDER BY "C1" DESC, "col2" NULLS LAST;')
@@ -137,7 +137,7 @@ describe('test orderBy Step', () => {
 
   it('Produces [SELECT "col1" AS " DESC" FROM "testTable" ORDER BY " DESC";]', () => {
     const actual = sql
-      .select(column1.as(' DESC'))
+      .select(col1.as(' DESC'))
       .from(table)
       .orderBy(' DESC')
       .getSQL()
@@ -147,7 +147,7 @@ describe('test orderBy Step', () => {
 
   it('Produces [SELECT "col1" AS " NULLS FIRST" FROM "testTable" ORDER BY " NULLS FIRST";]', () => {
     const actual = sql
-      .select(column1.as(' NULLS FIRST'))
+      .select(col1.as(' NULLS FIRST'))
       .from(table)
       .orderBy(' NULLS FIRST')
       .getSQL()