+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Changelog
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
+## [0.10.1]  2022-04-08
+### Changes
+- rename "column?" to "col?" in unit test
+- refactor group by aggregate functions
+- refactor all change getStmt arg to BuilderData
+- add Condition functions (eq, ne, gt...etc) to AggregateFunction
+- Move Expression into Expression.ts file
+- Move Condition into Condition.ts file
+- Move Operand class into Operand.ts file
+- Rename models to types and move it inside models folder
+
+### Added
+- Having step can contain aggregate function condition
+
 ## [0.10.0]  2022-03-26
 ### Changes
 - BaseStep is a class now that Step inherits from it